refactor: restructure directories under `example/`
Fixes https://github.com/nodekit-org/express-isomorphic/issues/8
