+++ example/launch.js
@@ -5,4 +5,4 @@ require('@babel/register')({
   extensions: ['.js', '.jsx', '.ts', '.tsx'],
 });
 
-require('./example.ts');
+require('./src/server/server.ts');
